move regions logic from waveformcontrol to audioplayerwithannotations
