+++ annotatedaudio/frontend/player/AudioPlayerWithAnnotation.svelte
@@ -1,11 +1,15 @@
 <script lang="ts">
-	import type { WaveformOptions } from "../shared/types";
+	import type { Annotation, WaveformOptions } from "../shared/types";
 	import type { I18nFormatter } from "@gradio/utils";
 
 	import { onMount } from "svelte";
-	import { Music } from "@gradio/icons";
+	import { Music, Undo} from "@gradio/icons";
+	import Gum from "../shared/icons/Gum.svelte";
 	import WaveSurfer from "wavesurfer.js";
-	import { skip_audio, process_audio } from "../shared/utils";
+	import RegionsPlugin, {
+		type Region
+	} from "wavesurfer.js/dist/plugins/regions.js";
+	import { skip_audio } from "../shared/utils";
 	import WaveformControls from "../shared/WaveformControls.svelte";
 	import { Empty } from "@gradio/atoms";
 	import { resolve_wasm_src } from "@gradio/wasm/svelte";
@@ -25,17 +29,25 @@
 
 	let container: HTMLDivElement;
 	let waveform: WaveSurfer | undefined;
-	let playing = false;
+	let wsRegions: RegionsPlugin;
+	let activeRegion: Region | null = null;
+	let leftRegionHandle: HTMLDivElement | null;
+	let rightRegionHandle: HTMLDivElement | null;
+	let activeHandle = "";
 
 	let timeRef: HTMLTimeElement;
 	let durationRef: HTMLTimeElement;
 	let audio_duration: number;
+	let playing = false;
 
 	let trimDuration = 0;
 
 	let show_volume_slider = false;
+	let showRedo = interactive;
 
-	let captionList = [];
+	let initialAnnotations: Annotation[] | null = null;
+	// correspondance between a Region and an Annotation
+	let regionsMap: Map<string, Annotation> = new Map();
 
 	const dispatch = createEventDispatcher<{
 		stop: undefined;
@@ -64,11 +76,145 @@
 		});
 	};
 
+	/**
+	 * Print annotations on waveform by linking regions to each annotations.
+	 * A region can be view as a visual representation of an annotation.
+	 */
+	const  addAnnotations = (): void =>{
+
+		var annotations = value.annotations;
+
+		// keep initial annotations in memory, for future retrieval
+		if (initialAnnotations === null){
+			initialAnnotations = []
+			annotations.forEach(
+				annotation => initialAnnotations.push(Object.assign({}, annotation))
+			);
+		}
+
+		value.annotations.forEach(annotation => {
+			var region = wsRegions.addRegion({
+				start: annotation.start,
+				end: annotation.end,
+				color: annotation.color,
+				drag: true,
+				resize: true,
+			})
+			region.element.style.top = (annotation.level * 10).toString() + "%";
+			region.element.style.height = (100 - (annotation.numLevels + 1) * 10 ).toString() + "%";
+
+			// link annotation to region
+			regionsMap.set(region.id, annotation);
+
+		});
+	}
+
+	/**
+	 * Remove the annotation linked to the specified region
+	 * @param region region of the annotation to be removed
+	 */
+	const removeAnnotation = (region: Region): void => {
+		regionsMap.delete(region.id)
+		region.remove();
+		updateAnnotations();
+	}
+
+	/**
+	 * Reset annotations to their initial state, ie annotations
+	 * provided by the pyannote pipeline
+	 */
+	const resetAnnotations = (): void  => {
+		clearAnnotations();
+		initialAnnotations.forEach(
+				annotation => value.annotations.push(Object.assign({}, annotation))
+		);
+		addAnnotations();
+		console.log(value)
+		dispatch("edit", value);
+	}
+
+	/**
+	 * update annotations with current regions' state
+	 */
+	const updateAnnotations = (): void => {
+		value.annotations = Array.from(regionsMap.values());
+		dispatch("edit", value);
+	}
+
+	/**
+	 * Clear all the annotations, and linked regions, from the waveform
+	 */
+	const clearAnnotations = (): void => {
+		wsRegions?.clearRegions();
+		value.annotations = []
+		regionsMap.clear()
+	};
+
+	/**
+	 * Select the annotation next (in terms of time) to current
+	 * active annotation. If active annotation is the last one,
+	 * the next region to be activated is the first annotation
+	 * on the waveform.
+	 */
+	const selectNextAnnotation = (shiftPressed: boolean): void => {
+		if(activeRegion !== null){
+			// do nothing if there is no active annotation
+			return;
+		}
+
+		// Go back if shift was pressed, else go ahead:
+		var direction = shiftPressed ? -1 : 1;
+		var regions = wsRegions.getRegions();
+		console.log(regions);
+		var activeRegionIdx = regions.indexOf(activeRegion);
+		activeRegion = regions.at((activeRegionIdx + direction) % regions.length);
+		console.log(activeRegion.id);
+	};
+
+	const adjustRegionHandles = (handle: string, key: string): void => {
+		let newStart;
+		let newEnd;
+
+		if (!activeRegion) return;
+		if (handle === "left") {
+			if (key === "ArrowLeft") {
+				newStart = activeRegion.start - 0.05;
+				newEnd = activeRegion.end;
+			} else {
+				newStart = activeRegion.start + 0.05;
+				newEnd = activeRegion.end;
+			}
+		} else {
+			if (key === "ArrowLeft") {
+				newStart = activeRegion.start;
+				newEnd = activeRegion.end - 0.05;
+			} else {
+				newStart = activeRegion.start;
+				newEnd = activeRegion.end + 0.05;
+			}
+		}
+
+		activeRegion.setOptions({
+			start: newStart,
+			end: newEnd
+		});
+
+		trimDuration = activeRegion.end - activeRegion.start;
+	};
+
 	$: if (container !== undefined) {
 		if (waveform !== undefined) waveform.destroy();
 		container.innerHTML = "";
 		create_waveform();
 		playing = false;
+	}
+
+	$: if(waveform){
+		wsRegions = waveform.registerPlugin(RegionsPlugin.create());
+	}
+
+	$: if(value?.annotations !== null && initialAnnotations === null){
+		addAnnotations()
 	}
 
 	$: waveform?.on("decode", (duration: any) => {
@@ -88,6 +234,14 @@
 		} else {
 			waveform?.play();
 		}
+		window.addEventListener("keydown", (e) => {
+			switch(e.key){
+				case "ArrowLeft":  adjustRegionHandles(activeHandle, "ArrowLeft"); break;
+				case "ArrowRight": adjustRegionHandles(activeHandle, "ArrowRight"); break;
+				case "Tab": selectNextAnnotation(e.shiftKey); break;
+				default: //do nothing
+			}
+		});
 	});
 
 	$: waveform?.on("finish", () => {
@@ -102,6 +256,55 @@
 		playing = true;
 		dispatch("play");
 	});
+
+	$: wsRegions?.on("region-updated", (region) => {
+		var updatedAnnotation = regionsMap.get(region.id);
+		updatedAnnotation.start = region.start;
+		updatedAnnotation.end = region.end;
+		updateAnnotations();
+	});
+
+	$: wsRegions?.on("region-clicked", (region, e) => {
+		e.stopPropagation(); // prevent triggering a click on the waveform
+		// if removal mode is enable, remove clicked region
+		if (region && mode === "remove") {
+			removeAnnotation(region)
+		}
+		else{
+			// update the active region
+			if(activeRegion !== null) {
+				activeRegion.element.classList.remove("active-region");
+			}
+			activeRegion = region;
+			console.log(activeRegion.id);
+			activeRegion.element.classList.add("active-region");
+			region.play();
+		}
+	});
+
+	$: if (activeRegion) {
+		const shadowRoot = container.children[0]!.shadowRoot!;
+
+		rightRegionHandle = shadowRoot.querySelector('[data-resize="right"]');
+		leftRegionHandle = shadowRoot.querySelector('[data-resize="left"]');
+
+		if (leftRegionHandle && rightRegionHandle) {
+			leftRegionHandle.setAttribute("role", "button");
+			rightRegionHandle.setAttribute("role", "button");
+			leftRegionHandle?.setAttribute("aria-label", "Drag to adjust start time");
+			rightRegionHandle?.setAttribute("aria-label", "Drag to adjust end time");
+			leftRegionHandle?.setAttribute("tabindex", "0");
+			rightRegionHandle?.setAttribute("tabindex", "0");
+
+			leftRegionHandle.addEventListener("focus", () => {
+				if (wsRegions) activeHandle = "left";
+			});
+
+			rightRegionHandle.addEventListener("focus", () => {
+				if (wsRegions) activeHandle = "right";
+			});
+		}
+	}
 
 	async function load_audio(data: string): Promise<void> {
 		await resolve_wasm_src(data).then((resolved_src) => {
@@ -155,22 +358,41 @@
 		</div>
 
 		{#if waveform}
-			<WaveformControls
-                value={value}
-				{container}
-				{waveform}
-				{playing}
-				{audio_duration}
-				{i18n}
-				{interactive}
-				bind:mode
-				bind:trimDuration
-				bind:show_volume_slider
-				showRedo={interactive}
-				{waveform_options}
-				{editable}
-				on:edit={(e) => dispatch("edit", e.detail)}
-			/>
+			<div class="commands">
+				<div class="waveform-controls">
+					<WaveformControls
+						{waveform}
+						{playing}
+						{audio_duration}
+						{i18n}
+						bind:show_volume_slider
+						{waveform_options}
+					/>
+				</div>
+				<div class="annotations-actions">
+					{#if editable && interactive && value.annotations}
+					{#if showRedo}
+						<button
+							class="action icon"
+							aria-label="Reset annotations"
+							title={i18n("Reset annotations")}
+							on:click={resetAnnotations}
+						>
+							<Undo/>
+						</button>
+					{/if}
+					<button
+						class="action icon remove-button"
+						aria-label="Remove an annotation"
+						title={i18n("Remove an annotation")}
+						on:focusin={() => mode = "remove"}
+						on:focusout={() => mode = ""}
+					>
+						<Gum/>
+					</button>
+				{/if}
+				</div>
+			</div>
 			{#if value?.annotations}
 				<Caption
 					value={value.annotations}
@@ -181,8 +403,38 @@
 {/if}
 
 <style>
+	.action {
+		width: var(--size-5);
+		width: var(--size-5);
+		color: var(--neutral-400);
+		margin-left: var(--spacing-md);
+	}
+
+	.annotations-actions {
+		display: flex;
+		justify-self: self-end;
+		align-items: center;
+	}
+
+	.commands {
+		display: flex;
+		justify-content: space-between;
+	}
+
 	.component-wrapper {
 		padding: 4em;
+	}
+
+	.icon:hover {
+		color: var(--color-accent);
+	}
+
+	.remove-button {
+		fill: #9ca3af;
+	}
+
+	.remove-button:hover, .remove-button:focus {
+		fill: var(--color-accent);
 	}
 
 	:global(::part(wrapper)) {
@@ -214,6 +466,10 @@
 		align-items: center;
 		justify-content: center;
 		width: var(--size-full);
+	}
+
+	.waveform-controls {
+		width: 80em;
 	}
 
 	#waveform {
